ai-trading-bot/requirements.txt
# ============================================================
# AI Trading Bot — Python Bağımlılıkları
# Python 3.11+ gereklidir
# Kurulum: pip install -r requirements.txt
# ============================================================

# ── Broker & Veri Kaynakları ───────────────────────────────
alpaca-py>=0.29.0
ccxt>=4.3.0
yfinance>=0.2.40
newsapi-python>=0.2.7
praw>=7.7.1
beautifulsoup4>=4.12.3
requests>=2.31.0
lxml>=5.0.0
feedparser>=6.0.10

# ── Veri İşleme & Feature Engineering ──────────────────────
pandas>=2.2.0
pandas-ta>=0.3.14b0
numpy>=1.26.4
scikit-learn>=1.4.0
scipy>=1.12.0

# ── Makine Öğrenimi Modelleri ──────────────────────────────
xgboost>=2.0.3
lightgbm>=4.3.0
joblib>=1.3.2

# ── AI / LLM ───────────────────────────────────────────────
anthropic>=0.29.0
openai>=1.30.0          # isteğe bağlı, alternatif LLM

# ── Bildirim ───────────────────────────────────────────────
python-telegram-bot>=21.3

# ── Görev Zamanlayıcı ──────────────────────────────────────
APScheduler>=3.10.4
pytz>=2024.1

# ── Dashboard & Görselleştirme ─────────────────────────────
streamlit>=1.35.0
plotly>=5.22.0

# ── Veritabanı ─────────────────────────────────────────────
sqlalchemy>=2.0.30

# ── Konfigürasyon ──────────────────────────────────────────
pyyaml>=6.0.1
python-dotenv>=1.0.0

# ── Yardımcı ───────────────────────────────────────────────
python-dateutil>=2.8.2
aiohttp>=3.9.0
tenacity>=8.2.3         # retry logic
structlog>=24.1.0       # structured logging
rich>=13.7.0            # güzel terminal çıktısı
